Modularize welcome page and use react routers
